Safe deletion of unused forum images with node-cron
Utilizing node-cron, we are now safely deleting orphaned images that haven't been submitted. The cron job is scheduled to run every 15 minutes, but only if the frontend NewThreadPage was last ineracted with >45 minutes ago, and the thread_id is NULL (meaning they haven't submitted).
This logic pings the API to see if there are any orphaned images to prune. If there aren't, it does nothing. If there are, they are safely deleted from uploads/forum_images. Simply focusing in on the markdown module marks you as actively interacting with the thread creation logic.